requirements.txt
streamlit>=1.55,<2
python-dotenv>=1.0.1,<2
groq>=0.13,<1
linkedin-api>=2.3,<3
httpx>=0.27,<1
